feat: create common build script
