+++ package.json
@@ -9,13 +9,12 @@
     "api:serve": "vitepress serve api",
     "docs:dev": "vitepress dev docs",
     "docs:build": "vitepress build docs",
-    "postdocs:build": "node ./docs/scripts/build.mjs",
     "docs:serve": "vitepress serve docs",
     "dev:dev": "vitepress dev dev",
     "dev:build": "vitepress build dev",
-    "postdev:build": "node ./dev/scripts/build.mjs",
     "dev:serve": "vitepress serve dev",
-    "build": "yarn api:build && yarn docs:build && yarn dev:build"
+    "build": "yarn api:build && yarn docs:build && yarn dev:build",
+    "postbuild": "node ./scripts/build.mjs"
   },
   "devDependencies": {
     "serve": "^14.2.1",
